Also rewrite GTDynamicResourcePack
